backend/requirements.txt
azure-functions
snowflake-connector-python
python-dotenv
requests
